route and pcx uninstall work - group rules next

+++ uninstall.sh
@@ -1,6 +1,7 @@
 #!/bin/bash
 
 #
+# Super-VPC
 # This uninstalls (DELETES!) everything.
 #
 
@@ -55,6 +56,13 @@ echo "OK... here we go..."
 
 VPCID=$(aws cloudformation list-exports --query "Exports[?Name=='$PREFIX-VpcId'].Value" --output text)
 echo "VPCID=$VPCID"
+
+if [[ $VPCID == "" ]]; then
+    echo "Could not find Stack related to $PREFIX ($REGION) - please double check PREFIX and Region."
+    exit 0
+fi
+
+
 # TODO - Count up interfaces - warn the user.
 #aws ec2 describe-network-interfaces --filter Name=vpc-id,Values=$VPCID
 #
@@ -87,8 +95,11 @@ aws s3 rm s3://$BUCKET --recursive
 WORKGROUP=$(aws athena list-work-groups --query "WorkGroups[?Description=='This workgroup has the queries related to vpc flow logs.'].Name" --output text)
 echo "WORKGROUP=$WORKGROUP"
 
-# This is like a force delete. Otherwise the CFN stack delete will fail...
-aws athena delete-work-group --work-group $WORKGROUP --recursive-delete-option
+if [[ $WORKGROUP != "" ]]; then
+    # This is like a force delete. Otherwise the CFN stack delete will fail...
+    aws athena delete-work-group --work-group $WORKGROUP --recursive-delete-option
+    sleep 10
+fi
 
 # Athena stack may not exist - that's OK
 STACK_NAME=$PREFIX-athena-query
@@ -106,115 +117,4 @@ echo "Deleting ($STACK_NAME) ..."
 aws cloudformation delete-stack --stack-name $STACK_NAME
 aws cloudformation wait stack-delete-complete --stack-name $STACK_NAME 
 
-
-exit
-
-
-# Find the Node Instance Role and remove any manually attached policies (otherwise stack delete will fail)
-NODE_INSTANCE_ROLE=$(aws cloudformation describe-stack-resources --stack-name "eksctl-$PREFIX-nodegroup-nodes" --query "StackResources[?LogicalResourceId=='NodeInstanceRole'].PhysicalResourceId" --output text)
-POLICIES=$(aws iam list-attached-role-policies --role-name $NODE_INSTANCE_ROLE --query "AttachedPolicies[*].PolicyArn" --output text )
-
-echo "Removing all policies attached to $NODE_INSTANCE_ROLE (To ensure a clean stack delete)"
-#
-# This is important because we might've added app specific policies, or CloudWatch, etc.
-#
-ARRAY=($POLICIES)
-for P in "${ARRAY[@]}"
-do
-    echo "Detaching $P..."
-    aws iam detach-role-policy --role-name $NODE_INSTANCE_ROLE --policy-arn $P
-done
-
-# Need to get all fargate profiles and loop to delete/remove
-FARGATE_PROFILES=$(aws eks list-fargate-profiles --cluster-name $PREFIX --query "fargateProfileNames[*]" --output text)
-echo "Looking for Fargate Profiles ..."
-ARRAY=($FARGATE_PROFILES)
-for F in "${ARRAY[@]}"
-do
-    echo "Profile: $F."
-    echo "Deleting Fargate Profile $F... (This can take awhile, be patient.)"
-    eksctl delete fargateprofile --cluster $PREFIX --name $F --wait 
-done
-
-# Find OIDC Provider
-# This gives us the https:// Issuer name
-OIDC_ISSUER=$(aws eks describe-cluster --name $PREFIX --query "cluster.identity.oidc.issuer" --output text)
-echo "OIDC_ISSUER=$OIDC_ISSUER.  Now we need to find the ARN..."
-
-echo "Checking all the existing ODIC providers to find the ARN of the one associated with the cluster..."
-# We'll just grab the ARN for now, and delete it later.
-OIDCS=$(aws iam list-open-id-connect-providers --query "OpenIDConnectProviderList[*].Arn" --output text)
-ARRAY_OIDCS=($OIDCS)
-OIDC_ARN=""
-for O in "${ARRAY_OIDCS[@]}"
-do
-    URL=$(aws iam get-open-id-connect-provider --open-id-connect-provider-arn $O --query "Url" --output text)
-    if [[ "https://$URL" == "$OIDC_ISSUER" ]]; then
-        OIDC_ARN=$O
-        echo "Found OIDC_ARN=$OIDC_ARN."
-    fi
-done 
-# OIDC_ARN should now be the ARN of the OIDC Provider that we'll delete later.
-
-STACK_NAME=eksctl-$PREFIX-addon-iamserviceaccount-kube-system-aws-load-balancer-controller
-echo "Deleting ($STACK_NAME) ..."
-aws cloudformation delete-stack --stack-name $STACK_NAME
-aws cloudformation wait stack-delete-complete --stack-name $STACK_NAME
-
-STACK_NAME=eksctl-$PREFIX-addon-iamserviceaccount-kube-system-cluster-autoscaler
-echo "Deleting ($STACK_NAME) ..."
-aws cloudformation delete-stack --stack-name $STACK_NAME
-aws cloudformation wait stack-delete-complete --stack-name $STACK_NAME
-
-STACK_NAME=eksctl-$PREFIX-addon-aws-ebs-csi-driver
-echo "Deleting ($STACK_NAME) ..."
-aws cloudformation delete-stack --stack-name $STACK_NAME
-aws cloudformation wait stack-delete-complete --stack-name $STACK_NAME
-
-STACK_NAME=eksctl-$PREFIX-nodegroup-nodes
-echo "Deleting ($STACK_NAME) ..."
-aws cloudformation delete-stack --stack-name $STACK_NAME
-aws cloudformation wait stack-delete-complete --stack-name $STACK_NAME
-
-STACK_NAME=eksctl-$PREFIX-addon-vpc-cni
-echo "Deleting ($STACK_NAME) ..."
-aws cloudformation delete-stack --stack-name $STACK_NAME
-aws cloudformation wait stack-delete-complete --stack-name $STACK_NAME
-
-# Find any "other" stacks that seem related.
-# If the stackname starts with "eksctl-CLUSTERNAME-" it's possibly related. But we ask to make sure.
-OTHER_STACKS=$(aws cloudformation list-stacks --query "StackSummaries[?StackStatus!='DELETE_COMPLETE'].StackName" --output text)
-echo "Looking for stacks that look related, but were created outside of the install script..."
-ARRAY=($OTHER_STACKS)
-for S in "${ARRAY[@]}"
-do
-    # Skip the -cluster stack - we'll delete that one AFTER any dependent stacks are gone.
-    if [[ $S != "eksctl-$PREFIX-cluster" ]]; then
-        if [[ "$S" == *"eksctl-$PREFIX-"* ]]; then
-            read -p "Stack named $S ($AWS_DEFAULT_REGION) appears related, but created manually. DO YOU WANT TO DELETE THIS STACK? (Yy) " -n 1 -r
-            echo    # (optional) move to a new line
-            if [[ $REPLY =~ ^[Yy]$ ]]; then
-                echo "Deleting stack $S..."
-                aws cloudformation delete-stack --stack-name $S
-                aws cloudformation wait stack-delete-complete --stack-name $S
-            else
-                echo "Leaving stack $S as-is..."
-            fi
-        fi
-    fi
-done
-
-STACK_NAME=eksctl-$PREFIX-cluster
-echo "Deleting ($STACK_NAME) ..."
-aws cloudformation delete-stack --stack-name $STACK_NAME
-aws cloudformation wait stack-delete-complete --stack-name $STACK_NAME
-
-
-if [[ $OIDC_ARN != "" ]]; then
-    echo "Deleting OIDC Provider $OIDC_ARN..."
-    aws iam delete-open-id-connect-provider --open-id-connect-provider-arn $OIDC_ARN
-fi
-	
-echo "Done."
-
-echo "You probably want to run ./99-uninstall.sh now, to remove the IAM roles and VPC."
+exit 0
